chore: create userProfile to display of of the work
